software/saxstat_gui_v1/requirements.txt
# SaxStat GUI v1 Requirements
# Python 3.8+

# GUI Framework
PyQt5>=5.15.0

# Serial Communication
pyserial>=3.5

# Data Processing
numpy>=1.21.0

# Plotting
pyqtgraph>=0.12.0

# Optional: Future enhancements
# openpyxl>=3.0.0     # Excel export
# pandas>=1.3.0        # Advanced data handling
# scipy>=1.7.0         # Scientific computing
# matplotlib>=3.5.0    # Alternative plotting
